Swap in the new waveform generators

+++ src/audio.h
@@ -8,6 +8,6 @@
 errcode queue_audio(const void* data, size_t nbytes);
 void pause_audio(bool pause);
 errcode init_audio();
-int queue_square_wave(unsigned tone_hz, unsigned duration_ms);
+int queue_square_wave(float tone_hz, unsigned duration_ms);
 
 #endif  // AUDIO_H
